+++ src/backend/core/api/serializers.py
@@ -17,7 +17,6 @@ from rest_framework import serializers
 
 from core import choices, enums, models, utils, validators
 from core.services import mime_types
-from core.services.ai_services import AI_ACTIONS
 from core.services.converter_services import (
     ConversionError,
     Converter,
@@ -792,33 +791,38 @@ class VersionFilterSerializer(serializers.Serializer):
     )
 
 
-class AITransformSerializer(serializers.Serializer):
-    """Serializer for AI transform requests."""
+class AIProxySerializer(serializers.Serializer):
+    """Serializer for AI proxy requests."""
 
-    action = serializers.ChoiceField(choices=AI_ACTIONS, required=True)
-    text = serializers.CharField(required=True)
-
-    def validate_text(self, value):
-        """Ensure the text field is not empty."""
-
-        if len(value.strip()) == 0:
-            raise serializers.ValidationError("Text field cannot be empty.")
-        return value
-
-
-class AITranslateSerializer(serializers.Serializer):
-    """Serializer for AI translate requests."""
-
-    language = serializers.ChoiceField(
-        choices=tuple(enums.ALL_LANGUAGES.items()), required=True
+    messages = serializers.ListField(
+        required=True,
+        child=serializers.DictField(
+            child=serializers.CharField(required=True),
+        ),
+        allow_empty=False,
     )
-    text = serializers.CharField(required=True)
+    model = serializers.CharField(required=True)
 
-    def validate_text(self, value):
-        """Ensure the text field is not empty."""
+    def validate_messages(self, messages):
+        """Validate messages structure."""
+        # Ensure each message has the required fields
+        for message in messages:
+            if (
+                not isinstance(message, dict)
+                or "role" not in message
+                or "content" not in message
+            ):
+                raise serializers.ValidationError(
+                    "Each message must have 'role' and 'content' fields"
+                )
 
-        if len(value.strip()) == 0:
-            raise serializers.ValidationError("Text field cannot be empty.")
+        return messages
+
+    def validate_model(self, value):
+        """Validate model value is the same than settings.AI_MODEL"""
+        if value != settings.AI_MODEL:
+            raise serializers.ValidationError(f"{value} is not a valid model")
+
         return value