+++ src/frontend/servers/y-provider/src/services/sentry.ts
@@ -1,12 +1,13 @@
 import * as Sentry from '@sentry/node';
 import { nodeProfilingIntegration } from '@sentry/profiling-node';
-
+import packageJson from '../../package.json';
 import { SENTRY_DSN } from '../env';
 
 Sentry.init({
   dsn: SENTRY_DSN,
   integrations: [nodeProfilingIntegration()],
-  tracesSampleRate: 0.1,
   profilesSampleRate: 1.0,
+  release: packageJson.version,
+  tracesSampleRate: 0.1,
 });
 Sentry.setTag('application', 'y-provider');